Global pathfinding using constant per map regions
The map is split into a fixed number of squared regions, each having a
distinguished centre. These centres are path targets for ants scouting
the map. Each region is timestamped, so regions visited farther in time
are visited more eagerly. However, the distance between the ant and its
centre is also taken into account. I.e. if more regions are timestamped
alike, the closer is chosen as the target.