Region waypoints introduced
A waypoint is used as a start/goal in interregional pathfinding (as
region centre was before). Waypoints are at first in regions centres.
However, as the map is revealed it may turn out that centres are
impassable - then waypoints are looked for in the same region. All path
searches betwen regions are then done on the newfound waypoint instead
of the impassable centre.